cleanup: internalize meta reader/writer implementations

+++ include/fstree.h
@@ -28,7 +28,7 @@ struct tree_xattr_t {
 	size_t num_attr;
 	size_t max_attr;
 
-	uint32_t block;
+	uint64_t block;
 	uint32_t offset;
 	uint32_t size;
